pyproject.toml
[tool.poetry]
name = "envsafe"
version = "0.1.1"
description = "Type-safe environment variables with automatic .env loading"
authors = ["Rajarshi Datta <rayan05rio@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "envsafe", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
ruff = "^0.4"
black = "^24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
